Gamestate and basic display added
We have a gamestate for the editor and some basic graphics.

I also wrote down my plan for making the actual editor.
Gonna make a modified bricks object that has a few new functions... Wish i made classes so that I could just inherit that stuff. Well, a lesson for the next game.